refactor: :recycle: add layout overlay
