+++ src/components/hooks/use-country-rule.tsx
@@ -24,8 +24,6 @@ export const useCountryRule = () => {
         is_row: !isEuDomain() && !isUkDomain(),
         is_dev: false,
         is_africa: false,
-        is_south_africa: false,
-        is_nigeria: false,
     })
 
     const { website_status } = useContext(DerivStore)
@@ -51,8 +49,6 @@ export const useCountryRule = () => {
         const is_uk_eu = !(!is_eu && !is_uk)
         const is_row = !is_uk_eu
         const is_dev = isLocalhost() || isTestlink()
-        const is_south_africa = user_ip_country === 'za'
-        const is_nigeria = user_ip_country === 'ng'
 
         if (website_status) {
             setRegion({
@@ -68,8 +64,6 @@ export const useCountryRule = () => {
                 is_uk_eu,
                 is_row,
                 is_dev,
-                is_nigeria,
-                is_south_africa,
             })
         }
     }, [residence, user_ip_country, website_status])
