+++ src/components/layout/index.ts
@@ -1,4 +1,3 @@
 import { LocationContext } from './location-context'
-import { CFDWarning } from './layout'
 
-export { LocationContext, CFDWarning }
+export { LocationContext }
