+++ src/components/layout/layout.tsx
@@ -15,6 +15,7 @@ import NavInterim from './nav/nav-interim'
 import NavSecurity from './nav/nav-security'
 import NavJumpIndice from './nav/nav-jump-indices'
 import Footer from './footer'
+import LayoutOverlay from './layout-overlay'
 import EURedirect, { useModal } from 'components/custom/_eu-redirect-modal'
 import { usePlatformQueryParam } from 'components/hooks/use-platform-query-param'
 import NonEuRedirectPopUp from 'components/custom/_non-eu-redirect-popup'
@@ -22,23 +23,12 @@ import { useCountryRule } from 'components/hooks/use-country-rule'
 import CookieBanner from 'components/custom/cookie-banner'
 import { CookieStorage } from 'common/storage'
 import { isBrowser, handleRedirect, isEuDomain } from 'common/utility'
-import { Localize } from 'components/localization'
-import { Text } from 'components/elements'
 import UKAccountClosureModal from 'components/layout/modal/uk_account_closure_modal'
-import device from 'themes/device'
 import { DerivStore, useDerivWS } from 'store'
-import { Container } from 'components/containers'
-import { loss_percent } from 'common/constants'
 import { usePageLoaded } from 'components/hooks/use-page-loaded'
 
 const LoadableFooter = Loadable(() => import('./footer'))
 const BeSquareFooter = Loadable(() => import('./besquare/footer'))
-const LiveChat = Loadable(() => import('./livechat'))
-const WhatsApp = Loadable(() => import('./whatsapp'))
-
-type CFDWarningProps = {
-    is_ppc: boolean
-}
 
 type LayoutProps = {
     children: ReactNode
@@ -46,7 +36,6 @@ type LayoutProps = {
     is_ppc?: boolean
     is_ppc_redirect?: boolean
     margin_top?: number | string
-    no_live_chat?: boolean
     no_login_signup?: boolean
     type?: string
 }
@@ -69,82 +58,6 @@ const has_dataLayer = isBrowser() && window.dataLayer
 const TRACKING_STATUS_KEY = 'tracking_status'
 const tracking_status_cookie = new CookieStorage(TRACKING_STATUS_KEY)
 
-const CFDWrapper = styled.section`
-    background-color: var(--color-grey-25);
-    background-size: cover;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    min-height: 7.4rem;
-    height: fit-content;
-    padding: 1.7rem 0 1.5rem;
-    position: fixed;
-    bottom: 0;
-    box-shadow: inset 0 1px 0 0 var(--color-grey-21);
-    z-index: 100;
-    @media (max-width: 826px) {
-        padding: 0.8rem 0;
-        height: 12.4rem;
-    }
-    @media (max-width: 710px) {
-        height: 10.8rem;
-    }
-    @media (max-width: 538px) {
-        height: 14rem;
-    }
-`
-
-const CFDContainer = styled(Container)`
-    @media ${device.bp1060} {
-        width: 90%;
-    }
-    @media ${device.tabletL} {
-        width: 95%;
-    }
-    @media ${device.tabletS} {
-        margin: 1rem auto;
-    }
-    @media ${device.mobileL} {
-        margin: 2rem auto;
-    }
-    @media ${device.mobileM} {
-        margin: 1rem auto;
-    }
-`
-
-const CFDText = styled(Text)`
-    font-size: 14px;
-
-    @media ${device.tablet} {
-        font-size: 12px;
-    }
-    @media ${device.mobileL} {
-        font-size: 10px;
-    }
-`
-
-export const CFDWarning = ({ is_ppc }: CFDWarningProps) => {
-    const { is_uk_eu } = useCountryRule()
-
-    if (is_ppc || is_uk_eu) {
-        return (
-            <CFDWrapper>
-                <CFDContainer>
-                    <CFDText>
-                        <Localize
-                            translate_text="CFDs are complex instruments and come with a high risk of losing money rapidly due to leverage. <0>{{loss_percent}}% of retail investor accounts lose money when trading CFDs with this provider.</0> You should consider whether you understand how CFDs work and whether you can afford to take the high risk of losing your money."
-                            values={{ loss_percent }}
-                            components={[<strong key={0} />]}
-                        />
-                    </CFDText>
-                </CFDContainer>
-            </CFDWrapper>
-        )
-    }
-    return <></>
-}
-
 const Main = styled.main<MainType>`
     margin-top: ${(props) => (props.margin_top && `${props.margin_top}rem`) || '7rem'};
     background: var(--color-white);
@@ -158,7 +71,6 @@ const Layout = ({
     is_ppc = false,
     is_ppc_redirect = false,
     margin_top = '',
-    no_live_chat = false,
     no_login_signup = false,
     type = '',
 }: LayoutProps) => {
@@ -254,7 +166,7 @@ const Layout = ({
             FooterNav = <Footer />
             break
         case 'static':
-            Navigation = <NavStatic is_ppc={is_ppc} />
+            Navigation = <NavStatic />
             break
         case 'interim':
             Navigation = <NavInterim interim_type={interim_type} />
@@ -313,45 +225,45 @@ const Layout = ({
         )
     }
     return (
-        <LocationProvider
-            has_mounted={is_mounted}
-            show_cookie_banner={show_cookie_banner}
-            toggleModal={toggleModal}
-            setModalPayload={setModalPayload}
-        >
-            {Navigation}
-            <Main margin_top={margin_top} is_static={is_static}>
-                {children}
-            </Main>
-            {show_cookie_banner && (
-                <CookieBanner
-                    onAccept={onAccept}
-                    onDecline={onDecline}
-                    is_open={show_cookie_banner}
+        <>
+            <LocationProvider
+                has_mounted={is_mounted}
+                show_cookie_banner={show_cookie_banner}
+                toggleModal={toggleModal}
+                setModalPayload={setModalPayload}
+            >
+                {Navigation}
+                <Main margin_top={margin_top} is_static={is_static}>
+                    {children}
+                </Main>
+                {show_cookie_banner && (
+                    <CookieBanner
+                        onAccept={onAccept}
+                        onDecline={onDecline}
+                        is_open={show_cookie_banner}
+                    />
+                )}
+                {FooterNav}
+                <EURedirect
+                    toggle={toggleModal}
+                    is_open={show_modal}
+                    closeModal={closeModal}
+                    to={modal_payload.to}
+                    target={modal_payload.target}
+                    rel={modal_payload.rel}
+                    ref={modal_payload.ref}
+                    aria_label={modal_payload.aria_label}
                 />
-            )}
-
-            {!no_live_chat && <LiveChat is_banner_shown={show_cookie_banner} />}
-            <WhatsApp />
-            {FooterNav}
-            <EURedirect
-                toggle={toggleModal}
-                is_open={show_modal}
-                closeModal={closeModal}
-                to={modal_payload.to}
-                target={modal_payload.target}
-                rel={modal_payload.rel}
-                ref={modal_payload.ref}
-                aria_label={modal_payload.aria_label}
-            />
-            <UKAccountClosureModal />
-            {show_non_eu_popup && (
-                <NonEuRedirectPopUp
-                    is_open={show_non_eu_popup}
-                    setShowNonEuPopup={setShowNonEuPopup}
-                />
-            )}
-        </LocationProvider>
+                <UKAccountClosureModal />
+                {show_non_eu_popup && (
+                    <NonEuRedirectPopUp
+                        is_open={show_non_eu_popup}
+                        setShowNonEuPopup={setShowNonEuPopup}
+                    />
+                )}
+            </LocationProvider>
+            <LayoutOverlay is_ppc={is_ppc} />
+        </>
     )
 }
