+++ src/components/layout/nav/nav-partner.tsx
@@ -1,6 +1,5 @@
 // Note: When using layout component for partners page, please add type='partners' and padding_top='10rem'
 import React from 'react'
-import { CFDWarning } from '../layout'
 import MainNav from './components/partner/main-nav'
 import NavPartnerMobile from './components/partner/nav-partner-mobile'
 import NavPartnerDesktop from './components/partner/nav-partner-desktop'
@@ -12,16 +11,13 @@ type NavPartnersProps = {
 
 const NavPartners = ({ hide_login_signup }: NavPartnersProps) => {
     return (
-        <>
-            <PartnerWrapper>
-                <MainNav />
-                <PartnerNavigationBarWrapper>
-                    <NavPartnerDesktop hide_login_signup={hide_login_signup} />
-                    <NavPartnerMobile hide_login_signup={hide_login_signup} />
-                </PartnerNavigationBarWrapper>
-            </PartnerWrapper>
-            <CFDWarning />
-        </>
+        <PartnerWrapper>
+            <MainNav />
+            <PartnerNavigationBarWrapper>
+                <NavPartnerDesktop hide_login_signup={hide_login_signup} />
+                <NavPartnerMobile hide_login_signup={hide_login_signup} />
+            </PartnerNavigationBarWrapper>
+        </PartnerWrapper>
     )
 }
